requirements.txt
torchsummary
torcheval
numpy==1.22.0
opencv-python
tqdm
tensorboard
wandb
idx2numpy
matplotlib==3.8